# sp cli: make sectors list much faster