## Related Issues
<!-- Link issues that this PR might resolve/fix. If an issue doesn't exist, include a brief motivation for the change being made -->

## Proposed Changes
Make sector checks run in parallel

## Additional Info
```
pre
./lotus-miner sectors list  31.82s user 1.64s system 59% cpu **56.525 total**
./lotus-miner sectors list --fast  34.44s user 1.97s system 329% cpu **11.060 total**


post
./lsm sectors list  15.59s user 1.72s system 248% cpu **6.971 total**
./lsm sectors list --fast  11.24s user 2.76s system 784% cpu **1.785 total**
```

## Checklist

Before you mark the PR ready for review, please make sure that:

- [ ] Commits have a clear commit message.
- [ ] PR title is in the form of of `<PR type>: <area>: <change being made>`
  - example: ` fix: mempool: Introduce a cache for valid signatures`
  - `PR type`: fix, feat, build, chore, ci, docs, perf, refactor, revert, style, test
  - `area`, e.g. api, chain, state, market, mempool, multisig, networking, paych, proving, sealing, wallet, deps
- [ ] New features have usage guidelines and / or documentation updates in
  - [ ] [Lotus Documentation](https://lotus.filecoin.io)
  - [ ] [Discussion Tutorials](https://github.com/filecoin-project/lotus/discussions/categories/tutorials)
- [ ] Tests exist for new functionality or change in behavior
- [ ] CI is green